requirements.txt
aiohttp
aiogram
beautifulsoup4
python-telegram-bot
requests
selenium
webdriver-manager








